refactor(suite): name the expandable group state and key by what they are
The keys are accountKey:groupType since the sell picker joined in, so
expandedAccountTokensGroups and updateExpandableAccountGroups were both
lies: useExpandableGroups now returns expandedGroupKeys and toggleGroup.

AssetGroupKey and getAssetGroupKey move out of sell's
buildGroupedAssetOptions into the picker kit, next to the AssetGroupOption
type they are derived from.
