+++ packages/suite/src/components/suite/asset-picker/hooks/index.ts
@@ -3,6 +3,6 @@ export * from './useListScrollReset';
 export * from './useSearchFilter';
 export * from './useInsertGroupLabelsAndSpaces';
 export * from './useFilterAccountsWithTokens';
-export * from './useExpandableAccountGroups';
+export * from './useExpandableGroups';
 export * from './useAccountsWithTokenDisplayNames';
 export * from './useTokenDisplaySymbolNames';
