+++ packages/suite/src/components/suite/asset-picker/utils/assetGroupKey.ts
@@ -1,0 +1,10 @@
+import { type AccountKey } from '@suite-common/wallet-types';
+
+import { type AssetGroupOption } from '../types';
+
+export type AssetGroupKey = `${AccountKey}:${AssetGroupOption['type']}`;
+
+export const getAssetGroupKey = (
+    accountKey: AccountKey,
+    groupType: AssetGroupOption['type'],
+): AssetGroupKey => `${accountKey}:${groupType}`;
