+++ packages/suite/src/views/wallet/send/Outputs/TokenSelect/SelectTokenAssetModal/SelectTokenAssetModal.tsx
@@ -16,7 +16,7 @@ import {
     ExpandableAssetRowGroup,
 } from 'src/components/suite/asset-picker/components';
 import {
-    useExpandableAccountGroups,
+    useExpandableGroups,
     useFilterAccountsWithTokens,
     useListScrollReset,
 } from 'src/components/suite/asset-picker/hooks';
@@ -59,8 +59,7 @@ export function SelectTokenAssetModal({
     const dispatch = useDispatch();
 
     const [search, setSearch] = useState('');
-    const { expandedAccountTokensGroups, updateExpandableAccountGroups } =
-        useExpandableAccountGroups();
+    const { expandedGroupKeys, toggleGroup } = useExpandableGroups();
     const listRef = useRef<HTMLDivElement>(null);
 
     const dataEnabled = getDefaultValue('options', []).includes('transactionData');
@@ -73,7 +72,7 @@ export function SelectTokenAssetModal({
 
     const options = useBuildTokenOptions({
         account,
-        expandedHiddenTokensGroups: expandedAccountTokensGroups,
+        expandedHiddenTokensGroups: expandedGroupKeys,
     });
     const filteredOptions = useFilterAccountsWithTokens(options, search);
 
@@ -185,7 +184,7 @@ export function SelectTokenAssetModal({
                             }
                             expanded={item.expanded}
                             onExpandToggle={expanded => {
-                                updateExpandableAccountGroups(item.account.key, expanded);
+                                toggleGroup(item.account.key, expanded);
                             }}
                             dataTestId={`@asset-picker/send-token/option/hidden-tokens/${item.account.symbol}`}
                         />
@@ -198,7 +197,7 @@ export function SelectTokenAssetModal({
                     return null;
             }
         },
-        [handleSelectChange, updateExpandableAccountGroups],
+        [handleSelectChange, toggleGroup],
     );
 
     return (
